requirements.txt
requests
python-dotenv
sigil-sdk
